fix: music loading
